requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.2
PyJWT==2.1.0
psycopg2-binary==2.9.1
